+++ packages/@expo/cli/src/export/__tests__/exportStaticAsync.test.ts
@@ -25,6 +25,7 @@ describe(getFilesToExportFromServerAsync, () => {
         },
         renderAsync,
         scripts: ['bar'],
+        cssLinks: ['baz'],
       })
     ).toEqual(
       new Map([